Add files via upload
This page contains all functions used in the paper "Two-Phase Sampling Designs for Data Validation in Settings with Covariate Measurement Error and Continuous Outcome". The paper compares different ways to select observations to have their records validated, aiming to reduce the variance of some parameter of interest. The designs are created for both likelihood and design-based estimators. For the former we applied multiple imputation and semiparametric maximum likelihood estimators and for the latter we applied inverse probability weighting and generalized raking.

 motivated by likelihood and design-based estimators to